refactor(providers): split auth & identity providers out of app_providers.dart
Seventh batch of #4506. Moves the 21-entity auth/identity cluster to
`lib/providers/auth_providers.dart`:

Foundational:
- secureKeyStorage, flutterSecureStorage, secureKeycastStorage
- oauthConfig, oauthClient
- pendingVerificationService, passwordResetListener, emailVerificationListener
- webAuthService, nostrKeyManager

Keystone:
- authService (the most-consumed provider in the codebase)
- currentAuthState, currentAuthRpcCapability (auth-stream bridges)
- knownAccounts (async list)

Identity adjuncts:
- nostrCreatorBindingServiceProvider
- cawgVerifierBaseUriProvider, cawgVerifierClientProvider
- zendeskIdentitySync (eager — app-shell-activated) + _setZendeskIdentity helper
- verifierClient, identityClaimsRepository
- nip98AuthService
- accountDeletionService

The three media-adjacent auth providers (blossomAuthService,
mediaViewerAuthServiceProvider, mediaAuthInterceptor) and the
_BlossomAuthAdapter private class deliberately STAY in
`app_providers.dart` — they're media-domain and will move with batch 9.

`app_providers.dart` adds both `import 'auth_providers.dart';` and
`export 'auth_providers.dart';`. The import is non-optional because 33+
internal sites consume `authServiceProvider`. The export keeps the 91+
external consumer files unchanged (largest consumer surface of any batch
so far).

Mutual import with `app_providers.dart` (auth_providers reads
`profileRepositoryProvider` for `_setZendeskIdentity`,
`userDataCleanupServiceProvider` for the `authService` factory). Dart
resolves library-level mutual imports lazily — same shape as
moderation_providers ↔ app_providers (batch 6).

Removed 17 now-unused imports from `app_providers.dart` via `dart fix`:
flutter_secure_storage, funnelcake_api_client, keycast_flutter,
nostr_key_manager, auth_rpc_capability model, environment_config model,
known_account model, account_deletion_service, cawg_verifier_client,
email_verification_listener, nip98_auth_service,
nostr_creator_binding_service, password_reset_listener,
pending_verification_service, web_auth_service, zendesk_support_service,
nostr_key_utils.

Codegen moved: 15+ *Provider chunks (~947 LoC) relocated from
`app_providers.g.dart` into `auth_providers.g.dart`.

Two follow-up findings (rolled into existing lists):
1. zendeskIdentitySync + _setZendeskIdentity is repository-layer
   coordination logic (auth-state-stream listener + 3-stage SDK plumbing)
   embedded in a provider. Joins the bridge-extraction follow-up.
2. _setZendeskIdentity inner JWT failure logged at Log.info; should be
   Log.warning per error_handling.md matrix for expected network/IO.

Outer Zendesk failure log level IS policy-compliant (Log.warning).

Refs #4506, #4339 Wave 2
